SDK-1344: Move Image into Yoti\Media namespace
